multiprocessing fix and refactor
